#57
added Speed, Volume, and weight to conversion classes. typecasted values as doubles to stop giving warnings

+++ app/src/main/kotlin/com/simplemobiletools/calculator/conversions/LengthConversion.kt
@@ -17,8 +17,8 @@ class LengthConversion:Converter {
     )
 
     override fun calculate(beginningQty: Double?, beginningUnitType: String, endingUnitType: String): Double {
-        var endingQty: Double = (1.0 / mapOfLengths.getValue(beginningUnitType))
-        endingQty *= mapOfLengths.getValue(endingUnitType)
+        var endingQty: Double = (1.0 / mapOfLengths.getValue(beginningUnitType) as Double)
+        endingQty *= mapOfLengths.getValue(endingUnitType) as Double
         if (beginningQty != null)
             endingQty *= beginningQty
         else
